requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alpaca-trade-api==3.2.0
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
deprecation==2.1.0
exceptiongroup==1.2.2
exchange_calendars==4.6
filelock==3.16.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
html5lib==1.1
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
korean-lunar-calendar==0.3.1
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
networkx==3.4.2
newsapi-python==0.2.7
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas_market_calendars==4.4.2
peewee==3.17.8
platformdirs==4.3.6
propcache==0.2.0
pyluach==2.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.1
ta==0.11.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toolz==1.0.0
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
yarl==1.18.0
yfinance==0.2.50
